change file structure to prevent importing standard libraries through faculty-distribtued
